eindversie 2020-07-24

+++ ZAPP/ZAPP/Library/ZorgmomentRecord.cs
@@ -32,7 +32,7 @@ namespace ZAPP
             opmerkingen = (string) record["opmerkingen"];
             aanwezigheid_begin = (string) record["aanwezigheid_begin"];
             aanwezigheid_eind = (string) record["aanwezigheid_eind"];
-            nieuw = (int)(Int32) record["nieuw"];
+            nieuw = (Int32) record["nieuw"];
         }
     }
